Adds an option to optimize the number of rows, cols
The number of cols and rows is determined by taking a sqrt of the number of key switches. So even if the layout is 14x3 keys = 42 keys. The number of GPIO pins used
will be 14, and the keys will be wired in a 7x7 matrix. As opposed to 17 GPIO pins if the layout col, row information is directly translated into connection matrix.